types to int to mend broken test

+++ landlab/components/flow_routing/lake_mapper.py
@@ -369,7 +369,7 @@ class DepressionFinderAndRouter(Component):
 
         # Record the number of pits and the IDs of pit nodes.
         self.number_of_pits = np.count_nonzero(self.is_pit)
-        self.pit_node_ids = np.where(self.is_pit)[0]
+        self.pit_node_ids = np.where(self.is_pit)[0].astype(int)
 
     def find_lowest_node_on_lake_perimeter(self, nodes_this_depression):
         """Locate the lowest node on the margin of the "lake".
@@ -625,7 +625,8 @@ class DepressionFinderAndRouter(Component):
                 pits = self._grid.at_node[pits]
                 supplied_pits = np.where(pits)[0]
                 self.pit_node_ids = np.setdiff1d(supplied_pits,
-                                                 self._grid.boundary_nodes)
+                                                 self._grid.boundary_nodes
+                                                 ).astype(int)
                 self.number_of_pits = self.pit_node_ids.size
                 self.is_pit.fill(False)
                 self.is_pit[self.pit_node_ids] = True
@@ -640,7 +641,8 @@ class DepressionFinderAndRouter(Component):
                 supplied_pits = pits
             # remove any boundary nodes from the supplied pit list
             self.pit_node_ids = np.setdiff1d(supplied_pits,
-                                             self._grid.boundary_nodes)
+                                             self._grid.boundary_nodes
+                                             ).astype(int)
             self.number_of_pits = self.pit_node_ids.size
             self.is_pit.fill(False)
             self.is_pit[self.pit_node_ids] = True
